Split out reporters #35
