+++ lib/attractor/reporters/html_reporter.rb
@@ -1,61 +1,6 @@
-# frozen_string_literal: true
-
-require 'descriptive_statistics/safe'
-require 'fileutils'
-require 'launchy'
-require 'tilt'
-
 module Attractor
-  # base reporter
-  class Reporter
-    extend Forwardable
-    attr_accessor :values
-    def_delegator :@watcher, :watch
-
-    def initialize(file_prefix: '')
-      @calculator = Calculator.new(file_prefix: file_prefix)
-      @values = @calculator.calculate
-      @suggester = Suggester.new(values)
-
-      @watcher = Watcher.new(file_prefix, lambda do
-        report
-      end)
-    end
-
-    def suggestions(quantile)
-      @suggestions = @suggester.suggest(quantile)
-      @suggestions
-    end
-
-    def report
-      @suggestions = @suggester.suggest
-    end
-
-    def render
-      'Attractor'
-    end
-  end
-
-  # console reporter
-  class ConsoleReporter < Reporter
-    def report
-      super
-      puts 'Calculated churn and complexity'
-      puts
-      puts "file_path#{' ' * 53}complexity   churn"
-      puts '-' * 80
-
-      puts @values.map(&:to_s)
-
-      puts
-      puts 'Suggestions for refactorings:'
-      @suggestions.each { |sug| puts sug.file_path }
-      puts
-    end
-  end
-
   # HTML reporter
-  class HtmlReporter < Reporter
+  class HtmlReporter < BaseReporter
     def report
       super
